pyproject.toml
[project]
name = "identity-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.18.4",
    "fastapi>=0.129.0",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.13.1",
    "pydantic[email]>=2.12.5",
    "python-jose[cryptography]>=3.5.0",
    "redis>=7.2.0",
    "slowapi>=0.1.9",
    "sqlalchemy>=2.0.46",
    "structlog>=25.5.0",
    "uvicorn[standard]>=0.41.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[project.scripts]
start = "uvicorn app.main:app"
dev = "uvicorn app.main:app --reload --host 0.0.0.0 --port 8000"
migrate = "alembic upgrade head"

[tool.coverage.run]
branch = true
source = ["app"]

[tool.coverage.report]
show_missing = true
skip_covered = true

[tool.ruff]
line-length = 100
target-version = "py314"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "C4"]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.14"
strict = true